+++ Task_0007_money_storage/money.h
@@ -3,23 +3,23 @@ struct money_value
 
 public:
 
-    money_value (int _int, int _hund);
+    money_value (long _int, long _hund);
     money_value (money_value const &other);
     money_value ();
     money_value &operator = (money_value const &other);
     money_value &operator += (money_value const &other);
     money_value &operator -= (money_value const &x);
-    money_value &operator *= (int x);
-    money_value &operator /= (int x);
-    money_value &operator %= (int x);
+    money_value &operator *= (long x);
+    money_value &operator /= (long x);
+    money_value &operator %= (long x);
     money_value operator + (money_value const &x) const;
     money_value operator - (money_value const &x) const;
-    money_value operator * (int x) const;
-    money_value operator / (int x) const;
-    money_value operator % (int x) const;
+    money_value operator * (long x) const;
+    money_value operator / (long x) const;
+    money_value operator % (long x) const;
     void print() const;
 
 private:
 
-    int value;
+    long value;
 };
